+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 #### 2.20.0 - 30.10.2015
 * Allow filtered updates of packages matching a regex - https://github.com/fsprojects/Paket/pull/1178
 * Search for paket.references in startup directory (auto-restore feature) - https://github.com/fsprojects/Paket/pull/1179
+* BUGFIX: Framework filtering for transisitve packages - https://github.com/fsprojects/Paket/issues/1182
 
 #### 2.19.0 - 29.10.2015
 * Resolver changed to breadth first search to escape more quickly from conflict situations - https://github.com/fsprojects/Paket/issues/1174